location accessablity added to shop.

+++ Models/Location.cs
@@ -12,10 +12,10 @@ namespace Zembil.Models
         public int LocationId { get; set; }
 
         [Required]
-        public int Longitude { get; set; }
+        public double Longitude { get; set; }
 
         [Required]
-        public int Latitude { get; set; }
+        public double Latitude { get; set; }
 
         [MaxLength(500)]
         public string LocationDescription { get; set; }
